requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
openai==1.12.0
Pillow==10.2.0
pandas==2.2.1
numpy==1.26.4
python-magic==0.4.27
requests==2.31.0
vapi-python==0.1.9
gumloop==0.1.0
temporalio==1.11.0
